Add files via upload
We would like to thank Dr. Hayley Hassler for her contribution as the author of the 2D scripts used to process the *.txt files and generate the 2D plots shown in Figure 4. These Python scripts were specifically designed to take the MASS-PRF output as input and create the visualizations depicted in the figure.